requirements.txt
python-telegram-bot==13.7
psycopg2-binary==2.9.1
python-dotenv==0.19.1
